first push

+++ src/main/java/cs455/overlay/wireformats/Protocol.java
@@ -1,9 +1,8 @@
 package cs455.overlay.wireformats;
 
 public interface Protocol {
-    byte REGISTER_REQ = 1;
+    int REGISTER_REQ = 1;
     byte REGISTER_RES = 2;
-
     byte DEREGISTER_REQ = 3;
     byte DEREGISTER_RES = 4;
 }
